Added basic user authentication and proper functionality gating. Working on adding file uploading for images.

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,16 @@ use App\Post;
 
 class PostsController extends Controller
 {
+        /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+        public function __construct()
+        {
+            $this->middleware('auth', ['except' =>['index', 'show']]);
+        }
+
     /**
      * Display a listing of the resource.
      *
@@ -46,10 +56,10 @@ class PostsController extends Controller
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
         $post->save();
         return redirect('/posts')->with('success', 'Post Created');
 
-        return 123;
     }
 
     /**
@@ -73,6 +83,13 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !==$post->user_id)
+        {
+            return redirect('/posts')->with('error', 'Unauthorized page.');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -86,8 +103,8 @@ class PostsController extends Controller
     public function update(Request $request, $id)
     {
         $this->validate($request, [
-        'title' => 'required',
-        'body' => 'required',
+            'title' => 'required',
+            'body' => 'required',
             ]);
 
         // Create Post
@@ -107,6 +124,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id !== $post->user_id)
+        {
+            return redirect('/posts')->with('error', 'Unauthorized page.');
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
